requirements.txt
pyrofork
tgcrypto-pyrofork
dnspython==2.6.1
Quart==0.19.6
PyYAML==6.0
regex
python-dotenv
motor
cachetools
requests
mail-parser==3.15.0